api holder moved where it cant hurt anyone. created card and search bar components and file structure
