+++ GoogleBooks/.gitignore
@@ -23,4 +23,4 @@ dist-ssr
 *.sln
 *.sw?
 
-./src/functions/tempAPIHolder.js
+/tempAPIHolder.js
