+++ GoogleBooks/src/components/app/App.jsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
-import { fetchBooks } from "./functions/helpers.js";
+import { fetchBooks } from "../../functions/helpers.js";
+import SearchBar from "../searchBar/SearchBar";
 
 function App() {
   return (
