+++ GoogleBooks/src/components/card/Card.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Card = () => {
+  return <div>Card</div>;
+};
+
+export default Card;
