+++ GoogleBooks/src/components/searchBar/SearchBar.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SearchBar = () => {
+  return <div>SearchBar</div>;
+};
+
+export default SearchBar;
